+++ src/main/java/pe/sia/service/interfaces/SoftwareService.java
@@ -6,9 +6,9 @@ import pe.sia.presentation.dto.activosDTO.SoftwareDTO;
 public interface SoftwareService {
 
     SoftwareDTO createSoftware(SoftwareDTO softwareDTO);
-    SoftwareDTO updateSoftware(Long idSoftware, SoftwareDTO softwareDTO);
-    void deleteSoftware(Long idSoftware);
-    SoftwareDTO getSoftwareById(Long idSoftware);
+    SoftwareDTO updateSoftware(Integer idSoftware, SoftwareDTO softwareDTO);
+    void deleteSoftware(Integer idSoftware);
+    SoftwareDTO getSoftwareById(Integer idSoftware);
     List<SoftwareDTO> getAllSoftwares();
     
 }
